requirements.txt
pefile
unpy2exe==0.4
uncompyle6==3.4.0
xdis
pycrypto
configparser
